shaders array turned into map
for conveinece and to prevent ordering accidents. it's a small point but
creating shaders when appending to the array has to be done in the
correct order. it's clearer and safer to use a map

renamed name of textureType type to shaderType

enabling and disabling of TV colour in the GUI now uses texture IDs for
a specially allocated texture, removing any possibility of accidentally
misuing a real texture ID