crt effect refinements
perlin noise no longer used as part of the interference effect. it's
still used to add a small amount of noise to help remove banding, but
it's always there

added YIQ interference

the aberration effect is also always appled. the difference with this
commit is that the when the effect is disabled, the intensity of the
effect is simply reduced to zero. this isn't the same as no applying
the effect at all

the aberration effect accumulates two pixels into one. previously an
average was taken but allowing the intensity to increase slightly is
effective and has been kept at 70% (rather than 50%)

black correction shader folded into crt effects shader

increased base saturation during colour generation

tweaked screen offset for solidstate bevel

increased magnitude of vignette effect