archivefs: set complete flag in all instances
if the name of the directory to be read hadn't changed then the result
was returned immediately but with the complete flag being set

this caused the ROM select window to forever centre on the currently
selected file. the bug was not noticeable unless the select window was
used for a second time in a session

the archivefs set channel now expects an instance of an Options type.
added a force option to instruct archivefs to always re-read directory