tidied up CRT related fragment shaders
including the removal of 'ghosting' shader. this was useful for
exagerrating the zookeeper brick effect but it wasn't necessary (because
of how we scale the image). also, it was a confusing name with regards
to the ghosting effect inherent in the chroma interference. rather than
change the name I just removed it

the crt 'enabled' setting has been renamed to 'pixel perfect' (which is
an inverted value). I find this easier to think about