changed colourburst value slightly for NTSC
I used a hue "value-1" as the basis for the phi calculation. however,
it makes much more sense to use "value" and then rotate by 33° (YUV to
YIQ conversion)

clarified (comments etc.) how the base phi value is adjusted before
taking the sin/cos, for both NTSC and PAL generation

change of phase slider in preferences window rounds the value to one
decimal place before saving to prefs value. this means that matching to
the preset value can be achieved