pixel presentation disables some rendering options if emulation is too slow
it is likely that CPU performance in the pause state can be improved,
particularly for non-active pauses. but we would need to be careful with
the colourgen package - we want changes to the colourgen prefs values to
show up on the screen even when paused. some sort of trigger from the
colourgen package would work but I'm not sure if the added complexity
is worth it