moved CRT prefrences from display to sdlimgui package
we needed more control over the post-change hooks for CRT values. in
particular we wanted the pixel perfect value to call the playscreen's
resize() function when changed. it would be awkward to do that from
a different package

incidentally, the resize() function needs to be called when toggling
pixel perfect mode becaus the CRT and pixel perfect screens may have
different scaling requirements