fixed DPC detection by adding the missing known size value
clarified through comments what the ideal size and the meaning of dumps
that are larger than the ideal size. (it is output from the cartridges
random number generator)

fixed RNG pump so that it more closely matches the description in the
DPC patent. clarfied that a reset value of 0xff for the RNG will produce
an endless sequence of 0xff values