corrected behaviour of POKE
for some reason now forgotten, my reasoning led me to think that POKEs should be
on read addresses rather than write addresses. that may well be correct
behaviour in some circumastances but the comment in the Poke function was not
clear about that

if it is correct in some circumstances to POKE a read address then we need to
think about a switch or another command completely

to be clear, my current thinking is that a POKE should replicate what the CPU
does when "writing" to a memory address. the PEEK command should replicate what
the CPU does when "reading" a memory address