added PokeNotify interface
because the TIA/RIOT addresses and the actual chip implementation are separated,
there was a flaw whereby a POKE to a TIA/RIOT address would not affect the chip
itself. the PokeNotify interface connects the memory to the chip for the
purposes of poking

the best solution is probably combining the TIA and RIOT implementations with
the memory functions. the separation was a very early design choice when I
didn't understand the 2600 very well and did not have a clear idea about how to
structure a Go program. but this is sufficient for now