moved faults package to coprocessor/ from coprocessor/developer
this is preparatory work for changes in the ARM package. while port the
ARM package to the test7800 project, I initially removed all references to the
developer package. however, the coprocessor package and the faults package were
still required. it makes sense therefore for the faults package to be moved here
too, even though we have the larger developer package