simplied arm T1 timer
ace/elf now instruct ARM to clock() during immediate mode. this is different
behaviour than that for cdf/dpc+ because of how and when yields occur for the
two different types of driver

I'm sure more work is required on this but it works well for now. good test ROMs:
    CDF - Draconian (console spec detection)
	- Boulderdash (console spec detection, speed of screen wipe and character
	  movement)
    ACE - MarcoJ's RPG (this uses the STM32's TIM2 rather T1)