fixed ARM emulation infiniteloop detection
I added this during ACE/ELF development but incorrectly set the yield type to
SyncWithVCS. this is fine for ACE/ELF but it caused serious issues for CDFJ and
DPC+, which don't respond to SyncWithVCS

the loop detection now returns a CycleLimit yield, which is more logically
correct

the infinite loop detection is very basic and only catches the pathological
instances where the PC loops back to the same address. in practice this means
C code of the form 'while (true) {}'. forcing a yield in this way was useful
during ELF development but is less useful now.

none-the-less, infinite loop detection is useful in principle so I've chosen to
keep the corrected code. moreover, an infinte loop will eventually result in a
CycleLimit yield so this is really just a quicker path to the same result

CDF and DPC+ also now explicitely assert that they never see a SyncWithVCS yield
and return with an ExecutionError yield if they do. ACE and ELF already
similarly asserted that they never see a ProgramEnded yield