fixed some festival/atarivox phonemes
the SayPhones() festival function does not work well with single
phonemes. for this reason, when single phonemes  are in the queue when
Flush() is called, the 'hh' phoneme is artificially added. this
additional phoneme is audible but it's better than not hearing anything
at all

for the same reason fast/slow and stress/relax have been clarified with
an explanatory comment as being disabled

the volume command is now implemented but it is only effective when the
volume is set to zero. when the volume value is zero, the festival
output is effectively disabled (there is no way using the SayPhones()
function to alter the volume of festival output)