moved handling of double-byte AtariVox commands to main atarivox code
the handling was part of the festival engine handler. this was
convenient when experimenting with the atarivox implementation but it is
not festival specific and so should be in the main driver code

added MSA tables from SpeakJet user manual. not all being used fully yet
but should be useful for providing contextual information to the user of
the debugger