careful queueing for frames for ffmpeg rendering
audio can get out-of-sync when rewinding. direct use of the wavwriter
package has been replaced by a shim in the video package. currently it
just forwards everything to wavwriter but there may be a better way of
controlling what is written to the wav such that rewinding is supported

a better option might be to prevent rewinding from happening when video
recording is active

fixed gl21 rendering