renamed functional_test package to klaus2m5
the new name indicates the name of github repository the tests are from

added decimal_mode test from klaus2m5 github repository. the functional_test,
which was already being used, has BCD tests of it's own but the
decimal_mode test can also optionally test the status flags after a BCD
operation

the test for the overflow flag currently remains disabled. this is
because the information in http://www.6502.org/tutorials/decimal_mode.html,
used to implement BCD in Gopher2600, does not seem to agree with the tests