gl32 render done in two passes
the first pass is just the playmode screen and bevel (and bevel rim).
after this first pass the screenshot and video processing is performed.
the second pass adds the windows, overlay, etc.

if a screenshot is made during video recording then the screen dimming,
caused by the lack of interference during screenshotting, is visible in
the video

these changes have not been replicated for gl21 rendering