moved subtitling to it's own file (away from the overlay)
this was intended to help allow video recording of subtitles. however,
the current method of eliding elements from the video recording is based
on textures. the subtitling is still drawn alongside the other windows
in the imgui texture so cannot be isolated in that way

the solution would be to draw the subtitling to it's own texture and
treat it like the main playscreen textures for recording purposes