text input supression when ` key is pressed
the ` key is used to switch between debugger and playmode immediately.
if the debugger terminal was open for example, switching to playmode
would cause the ` symbol to appear in the terminal input

the downside of this is that it's impossible to use the ` symbol for
anything. we could perhaps have an exception for the suppression rule
that allows the use of ` when the specifying the filename for a ROM